api/requirements.txt
Django==3.1
djangorestframework==3.11.1
django-rest-auth==0.9.5
django-cors-headers==3.5.0
psycopg2-binary==2.8.5
pandas==1.1.2
numpy==1.19.5
tensorflow-cpu==2.6.0
gensim==4.0.0
joblib==0.16.0
gunicorn==20.0.4
pytz==2017.3
pycocotools==2.0.1
underthesea